Truncation errors resolved

+++ Position.cpp
@@ -1,11 +1,11 @@
 #include "Position.h"
 
-Position::Position(float x, float y) {
+Position::Position(int x, int y) {
 	pos[0] = x;
 	pos[1] = y;
 }
 
-void Position::set(float x, float y) {
+void Position::set(int x, int y) {
 	pos[0] += x;
 	pos[1] += y;
 }
